backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
pydantic==2.9.0
pydantic-settings==2.6.0
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
asyncio-mqtt==0.16.1
httpx==0.27.2
aiomqtt>=1.0.0

aiosqlite==0.20.0
